refactor: Remove index domain file, rename chart-id, fix Tooltip type vs. component
