+++ app/components/charts-generic/use-interaction.tsx
@@ -6,7 +6,7 @@ import {
   useContext,
   useReducer,
 } from "react";
-import { Observation } from "../../domain";
+import { Observation } from "../../domain/data";
 
 export interface InteractionElement {
   visible: boolean;
