--- app/domain/index.ts
@@ -1,2 +1,0 @@
-export * from "./data";
-export * from "./chart-id";
