+++ app/graphql/resolvers.ts
@@ -9,7 +9,7 @@ import { descending } from "d3-array";
 import fuzzaldrin from "fuzzaldrin-plus";
 import { GraphQLJSONObject } from "graphql-type-json";
 import { Filters } from "../configurator";
-import { parseObservationValue } from "../domain";
+import { parseObservationValue } from "../domain/data";
 import { SPARQL_ENDPOINT } from "../domain/env";
 import { locales, parseLocaleString } from "../locales/locales";
 import {
