+++ app/lib/create-chart-id.ts
